% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{character_vector_as_def_string}
\alias{character_vector_as_def_string}
\title{Convert Character Vector to Defining String}
\usage{
character_vector_as_def_string(pvec)
}
\arguments{
\item{pvec}{vector to be converted to definition string}
}
\value{
s_result defining string for pvec
}
\description{
Given a character vector, we want to return  the string
that defines the same vector using 'c()'
}
\details{
This is needed when producing R-statments for
exercises and exams
}
