% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_function.R
\name{create_pedigree}
\alias{create_pedigree}
\title{Create Pedigree Using AlphaSimR}
\usage{
create_pedigree(
  pn_nr_founder,
  pn_nr_cross_f1,
  pn_nr_gen,
  pvec_nr_cross,
  ps_set_sex = "yes_sys"
)
}
\arguments{
\item{pn_nr_founder}{number of founder animals (animals without parents)}

\item{pn_nr_cross_f1}{nuber of animals in f1 (offspring of founders)}

\item{pn_nr_gen}{number of generation to simulate}

\item{pvec_nr_cross}{the number of crosses in each generation}

\item{ps_set_sex}{how sex should be set in AlphaSimR}
}
\value{
pedigree as tibble
}
\description{
Pedigree creation with AlphaSimR is wrapped into this function
}
\details{
\subsection{title: Pedigree Creation Functions
date:  2020-12-16}{
}
}
