% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_function.R
\name{exam2rmd}
\alias{exam2rmd}
\title{Generate Rmd Solutions and Exams}
\usage{
exam2rmd(
  pvec_problem,
  pvec_names = NULL,
  pn_nr_exam = 1,
  ps_rmd_dir = "rmd",
  ps_rsrc_dir = "R",
  ps_tex_dir = "tex",
  ps_outdir = "out",
  pb_force = FALSE
)
}
\arguments{
\item{pvec_problem}{vector of problem names}

\item{pvec_names}{vector of student names}

\item{pn_nr_exam}{number of exams to be generated, if pvec_names is missing}

\item{ps_rmd_dir}{directory where rmd-sources for the problems are stored}

\item{ps_rsrc_dir}{directory where R-scripts for problem-setup are stored}

\item{ps_tex_dir}{directory containing additional tex files}

\item{ps_outdir}{directory where output should be stored}

\item{pb_force}{flag to first delete old output directories}
}
\description{
Given a vector of problem names and a vector of student names, a series of
individualised solutions and exams are created. The output currently consists
of Rmarkdown source files and pdf-files for the exam questions and the solutions.
Each problem requires a Rmarkdown source file which is stored in a file named
the same way as the problem name. Hence the source of the problem 'nrm' must
be stored in a file called 'nrm.Rmd'. Any problem setup should be included in
an R-script called 'nrm.R' which contains a single function 'generate_problem_setup()'
}
\details{
---
title: Create Rmarkdown Exam
date:  2020-12-15
---


Solutions to questions must be tagged with the begin-tag '\solstart' and the
end-tag '\solend'. These tags have to be defined in the yaml header or in a
preamble.tex file
}
