% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_function.R
\name{extend_pedigree}
\alias{extend_pedigree}
\title{Extend Pedigree}
\usage{
extend_pedigree(ptbl_ped)
}
\arguments{
\item{ptbl_ped}{input pedigree to be extended by founders}
}
\value{
pedigree extended by founder animals
}
\description{
A given pedigree is extended with its founder animals.
}
\details{
The founder animals are determined by taking the set-difference between
the vector of parents and the vector of animal IDs
}
