% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{numeric_vector_as_def_string}
\alias{numeric_vector_as_def_string}
\title{Convert Numeric Vector to Defining String}
\usage{
numeric_vector_as_def_string(pvec)
}
\arguments{
\item{pvec}{vector to be converted to definition string}
}
\value{
s_result defining string for pvec
}
\description{
Given a numeric vector, we want to return  the string
that defines the same vector using 'c()'
}
\details{
\subsection{title: Helper Functions For Exam Generation
date:  2020-12-15}{

This is needed when producing R-statments for
exercises and exams
}
}
