% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_function.R
\name{create_pedigree}
\alias{create_pedigree}
\title{Create Pedigree Using AlphaSimR}
\usage{
create_pedigree(
  pn_nr_founder,
  pn_nr_cross_f1,
  pn_nr_gen,
  pvec_nr_cross,
  ps_set_sex = "yes_sys"
)
}
\value{
pedigree as tibble
}
\description{
Pedigree creation with AlphaSimR is wrapped into this function
}
\details{
\subsection{title: Pedigree Creation Functions
date:  2020-12-16}{
}
}
